requirements.txt
autoflake
colorama==0.3.7
Cython==0.27.3
future
ipython
isort
loguru==0.4.1
matplotlib
numba==0.45.0
numpy==1.16.0
opencv-python
pandas
pillow
pycocotools
pylint
scikit-image
scipy==1.2.1
shapely==1.7.0
tensorboard
tqdm
wget
yacs
yapf==0.28.0
pre-commit
